update Repo thuộc tính

+++ src/main/java/com/example/datn/repository/ThuocTinhRepo/CheDoGioRepo.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-public interface CheDoGioRepo extends JpaRepository<CheDoGio, Long> {
+public interface CheDoGioRepo extends JpaRepository<CheDoGio, Integer> {
     @Query("SELECT cdg FROM CheDoGio cdg " +
             "WHERE (LOWER(cdg.ten) LIKE LOWER(CONCAT('%', :query, '%')) " +
             "AND (:trangThai IS NULL OR cdg.trangThai = :trangThai))")
